requirements.txt
firecrawl-py
html2text
pydantic
bs4
emoji
python-dotenv
openai
tiktoken